Provide a mock of SugarCRM, so you can work on this app offline
If you are sat on a plane for 10 hours, and want to work on this
application, you are going to have an issue. Unless you have a virtual
machine with SugarCRM running locally. But then you have to be worried
about your battery!! Running the full SugarCRM application is overkill,
so now you can run a lightweight mock that returns "canned data". It
accepts any credentials, and returns you a record for each module type.